+++ app.vue
@@ -6,4 +6,5 @@
 </template>
 
 <script setup lang="ts">
+import NotificationsProvider from '~/project/common/components/NotificationsProvider.vue'
 </script>
