+++ pages/index.vue
@@ -4,7 +4,7 @@
 
 <script setup lang="ts">
 import { useScannersStore } from '~/stores/scanners'
-import SvgLoader from '~/common/components/SvgLoader.vue'
+import SvgLoader from '~/project/common/components/SvgLoader.vue'
 
 const scannersStore = useScannersStore()
 const scanners = scannersStore.scanners
