+++ pages/scan/list.vue
@@ -40,9 +40,9 @@
 
 <script lang="ts">
 import { mapState } from 'pinia'
-import ScanList from '~/project/scans/componenets/scanList.vue'
+import ScanList from '~/project/scans/components/scanList.vue'
 import { DfBreadcrumbs } from '~/dragonfly/components/Sections/DfBreadcrumbs'
-import SvgLoader from '~/common/components/SvgLoader.vue'
+import SvgLoader from '~/project/common/components/SvgLoader.vue'
 import type { VulnerabilityDetailBreadcrumbsType } from '~/dragonfly/components/Sections/DfBreadcrumbs/types'
 
 interface Data {
