+++ project/common/components/LoadingDialog.vue
