+++ project/common/components/NavigationAppBar.vue
