+++ project/common/components/NavigationDrawer.vue
