+++ project/common/components/SvgLoader.vue
