+++ project/scanners/components/ScannerSelect.vue
