+++ project/scans/components/VulnzTable.vue
