+++ project/scans/components/form/CreateNetworkScanForm.vue
@@ -1,8 +1,8 @@
 <template>
   <v-stepper-vertical-item
-    title="Target URLs / domains"
+    title="Target IPs"
     subtitle="required"
-    :error="rawUrlsErrorMessages.length > 0"
+    :error="ipErrorMessages.length > 0"
     :value="step"
   >
     <LoadingDialog
@@ -11,24 +11,26 @@
     />
     <v-form v-model="isFormValid">
       <p class="subtitle-1 mb-3">
-        Please specify the target URLs / domains:
+        Please specify the target IPs:
       </p>
       <v-textarea
-        v-model="rawUrls"
-        :error-messages="rawUrlsErrorMessages"
+        v-model="ip"
+        :error-messages="ipErrorMessages"
         variant="outlined"
         density="compact"
         clearable
-        label="URLs / domains"
-        placeholder="Add URLs and domains each on a separate line"
-        prepend-icon="mdi-web"
+        label="IP (IPv4 or IPv6)"
+        placeholder="Add IPs each on a separate line"
+        prepend-icon="mdi-ip-network-outline"
+        @input="addToTargetIps"
+        @keyup.enter="addToTargetIps"
       />
     </v-form>
     <template #next="{ next }">
       <v-btn
         color="primary"
         variant="elevated"
-        :disabled="isContinueDisabled"
+        :disabled="isFormValid === false"
         @click="next"
       >
         <v-icon start>
@@ -59,19 +61,19 @@
 
 <script lang="ts">
 import validator from 'validator'
-import LoadingDialog from '~/common/components/LoadingDialog.vue'
+import LoadingDialog from '~/project/common/components/LoadingDialog.vue'
 import type { Scanner } from '~/project/types'
-import AgentGroupSelect from '~/scan/components/AgentGroupSelect.vue'
+import AgentGroupSelect from '~/project/scans/components/AgentGroupSelect.vue'
 
 interface Data {
-  rawUrls: string | null
+  ip: string | null
+  userIps: Array<string>
   loadingDialog: boolean
-  isFormValid: boolean
   selectedAgentGroup: unknown
 }
 
 export default defineComponent({
-  name: 'CreateWebScanForm',
+  name: 'CreateNetworkScanForm',
   components: {
     LoadingDialog,
     AgentGroupSelect
@@ -93,41 +95,44 @@ export default defineComponent({
   emits: ['reset', 'update:assets', 'update:agent-group-id', 'createScan'],
   data(): Data {
     return {
-      isFormValid: false,
-      rawUrls: null,
+      ip: null,
+      userIps: [],
       loadingDialog: false,
       selectedAgentGroup: null
     }
   },
   computed: {
     /**
-     * The formatted user URLs.
+     * The errors for the IP messages.
      */
-    userUrls(): Array<string> {
-      return this.rawUrls?.split('\n').filter(Boolean) || []
-    },
-    /**
-     * The error messages for the user URLs.
-     */
-    rawUrlsErrorMessages(): Array<string> {
+    ipErrorMessages(): Array<string> {
       const errors: Array<string> = []
-      for (const url of this.userUrls) {
-        if (validator.isURL(url) === false) {
-          errors.push(`URL ${url} is invalid`)
+      if (this.userIps === undefined) {
+        return errors
+      }
+      for (const userIp of this.userIps) {
+        if (this.isValidIpOrIpRange(userIp) === false) {
+          errors.push(`IP or IP Range is invalid: ${userIp}`)
         }
       }
       return errors
     },
-    isContinueDisabled(): boolean {
-      return this.rawUrlsErrorMessages.length > 0 || this.rawUrls === null || this.rawUrls?.trim() === ''
+    /**
+     * Whether the form is valid or not.
+     */
+    isFormValid(): boolean | undefined {
+      if (this.userIps === undefined) {
+        return
+      }
+      return this.userIps?.length > 0 && this.ipErrorMessages.length === 0
     }
   },
   watch: {
-    userUrls: {
+    userIps: {
       deep: true,
       immediate: false,
       handler(newVal) {
-        this.$emit('update:assets', [{ url: { links: newVal } }])
+        this.$emit('update:assets', [{ network: { networks: newVal } }])
       }
     },
     selectedAgentGroup: {
@@ -139,15 +144,43 @@ export default defineComponent({
   },
   methods: {
     /**
-     * Clear the input.
+     * Check if an IP or IP range is valid.
+     * @param lastAddedIp
+     */
+    isValidIpOrIpRange(lastAddedIp: string): boolean | undefined {
+      if (this.ip === null || this.ip === undefined) {
+        return
+      }
+      if (validator.isIP(lastAddedIp) === true || validator.isIPRange(lastAddedIp) === true) {
+        return true
+      }
+      return false
+    },
+    /**
+     * Add an IP to a list of IPs.
+     */
+    addToTargetIps(): void {
+      this.userIps = this.ip?.split('\n').filter(Boolean) || []
+      if (this.userIps === undefined) {
+        this.userIps = []
+        return
+      }
+      const lastAddedIp = this.userIps[this.userIps?.length - 1]
+      if (lastAddedIp !== undefined && this.isValidIpOrIpRange(lastAddedIp) === true) {
+        this.userIps.push(lastAddedIp)
+      }
+    },
+    /**
+     * Clear the user input.
      */
     clear(): void {
-      this.rawUrls = null
+      this.ip = null
+      this.userIps = []
     },
     /**
      * Create a scan.
      */
-    createScan() {}
+    createScan(): void {}
   }
 })
 </script>
