+++ project/scans/components/form/CreateScanForm.vue
@@ -126,16 +126,15 @@
 
 <script lang="ts">
 import { mapActions } from 'pinia'
-import type { Scanner } from '~/project/types'
-import { AssetEnum, type Group } from '~/scan/types'
-import AssetTypeSelector from '~/scan/components/AssetTypeSelector.vue'
-import CreateMobileScanStoreForm from '~/scan/components/form/CreateMobileScanStoreForm.vue'
-import CreateWebScanForm from '~/scan/components/form/CreateWebScanForm.vue'
-import CreateWebApiScanForm from '~/scan/components/form/CreateWebApiScanForm.vue'
-import CreateNetworkScanForm from '~/scan/components/form/CreateNetworkScanForm.vue'
-import CreateMobileScanFileForm from '~/scan/components/form/CreateMobileScanFileForm.vue'
-import CreateScanYamlForm from '~/scan/components/form/CreateScanYamlForm.vue'
-import ScannerSelect from '~/scan/components/ScannerSelect.vue'
+import { AssetEnum, type Group, type Scanner } from '~/project/types'
+import AssetTypeSelector from '~/project/scans/components/AssetTypeSelector.vue'
+import CreateMobileScanStoreForm from '~/project/scans/components/form/CreateMobileScanStoreForm.vue'
+import CreateWebScanForm from '~/project/scans/components/form/CreateWebScanForm.vue'
+import CreateWebApiScanForm from '~/project/scans/components/form/CreateWebApiScanForm.vue'
+import CreateNetworkScanForm from '~/project/scans/components/form/CreateNetworkScanForm.vue'
+import CreateMobileScanFileForm from '~/project/scans/components/form/CreateMobileScanFileForm.vue'
+import CreateScanYamlForm from '~/project/scans/components/form/CreateScanYamlForm.vue'
+import ScannerSelect from '~/project/scanners/components/ScannerSelect.vue'
 import AssetsService from '~/project/assets/services/assets.service'
 import ScansService from '~/project/scans/services/ScanService'
 import { useNotificationsStore } from '~/stores/notifications'
