+++ project/scans/components/vulnz/VulnerabilityDetail.vue
@@ -91,7 +91,7 @@ import { mapActions } from 'pinia'
 import { Mutex } from 'async-mutex'
 import { sort } from 'fast-sort'
 import { useNotificationsStore } from '~/stores/notifications'
-import VulnerabilityTitle from '~/scan/components/vulnz/VulnerabilityTitle.vue'
+import VulnerabilityTitle from '~/project/scans/components/vulnz/VulnerabilityTitle.vue'
 import { DfMaterialCard } from '~/dragonfly/components/Cards/DfMaterialCard'
 import { DfBreadcrumbs } from '~/dragonfly/components/Sections/DfBreadcrumbs'
 import { DfEnhancedMdRender } from '~/dragonfly/components/Sections/DfMdRender'
